requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.7.0
rich>=13.0.0
tqdm>=4.65.0 